requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
pytz==2024.2
requests==2.32.3
urllib3==2.2.3
waveshare-epaper==1.3.0
gpiozero==2.0.1
lgpio==0.2.2.0
